+++ package.json
@@ -23,6 +23,7 @@
     "dependencies": {
         "@hookform/resolvers": "^3.1.1",
         "@react-native-async-storage/async-storage": "^1.18.2",
+        "@react-native-clipboard/clipboard": "^1.11.2",
         "@react-navigation/bottom-tabs": "^6.5.7",
         "@react-navigation/native": "^6.1.6",
         "@react-navigation/native-stack": "^6.9.12",
