+++ src/packages/uikit/components/AppBar/index.ts
@@ -1,0 +1,3 @@
+import AppBar from './AppBar';
+
+export default AppBar;
