+++ src/packages/uikit/components/Drawer/index.ts
@@ -1,0 +1,3 @@
+import Drawer from './Drawer';
+
+export default Drawer;
