+++ src/packages/uikit/components/EmptyListView/index.ts
@@ -1,0 +1,3 @@
+import EmptyListView from './EmptyListView';
+
+export default EmptyListView;
