+++ src/packages/uikit/components/ListView/index.ts
@@ -1,0 +1,3 @@
+import ListView from './ListView';
+
+export default ListView;
