+++ src/packages/uikit/components/Loading/index.ts
@@ -1,0 +1,3 @@
+import Loading from './Loading';
+
+export default Loading;
