+++ src/packages/uikit/components/OtpInput/index.ts
@@ -1,0 +1,3 @@
+import OTPInput from './OtpInput';
+
+export default OTPInput;
